Video Labels/requirements.txt
google-generativeai==0.7.2
pytest==8.3.3 